[occm] Remove volume related code from openstack-cloud-controller-manager
