+++ pkg/cloudprovider/providers/openstack/openstack_client.go
@@ -45,28 +45,6 @@ func (os *OpenStack) NewComputeV2() (*gophercloud.ServiceClient, error) {
 	return compute, nil
 }
 
-// NewBlockStorageV1 creates a ServiceClient that may be used with the Cinder v1 API
-func (os *OpenStack) NewBlockStorageV1() (*gophercloud.ServiceClient, error) {
-	storage, err := openstack.NewBlockStorageV1(os.provider, gophercloud.EndpointOpts{
-		Region: os.region,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize cinder v1 client for region %s: %v", os.region, err)
-	}
-	return storage, nil
-}
-
-// NewBlockStorageV2 creates a ServiceClient that may be used with the Cinder v2 API
-func (os *OpenStack) NewBlockStorageV2() (*gophercloud.ServiceClient, error) {
-	storage, err := openstack.NewBlockStorageV2(os.provider, gophercloud.EndpointOpts{
-		Region: os.region,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize cinder v2 client for region %s: %v", os.region, err)
-	}
-	return storage, nil
-}
-
 // NewBlockStorageV3 creates a ServiceClient that may be used with the Cinder v3 API
 func (os *OpenStack) NewBlockStorageV3() (*gophercloud.ServiceClient, error) {
 	storage, err := openstack.NewBlockStorageV3(os.provider, gophercloud.EndpointOpts{
